requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
supabase==2.7.4
google-generativeai==0.8.3
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
scikit-learn==1.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
httpx==0.27.2
streamlit==1.39.0
